Shared Mixer refactored to use raw pointers instead of shared pointers. This refactor impacted MegaDrone and RhythmGame.
This prevents ambiguous ownership of audio components, leaving the owner of the Mixer as the sole owner of the inputs to the Mixer.
This allows for objects with the Mixer as a member along with its inputs to control their duration automatically, in aggregate.
This fixes crashes on destruction of objects composing the Mixer and various inputs. (E.g. Fixes #521).