OboeTester: fixed crash when exiting Auto Glitch Test
Hitting Back button while in the middle of the
Auto Glitch Test could cause a crash.

I was calling closeAudio when the test thread finished.
AND I was calling closeAudio() from the Activity onStop() call.
They were colliding as I deleted the Oboe Stream.
So I synchronized the closeAudio() method and protected it with a new CLOSING state.