tests: Don't ignore perfMode in TestStreamWaitState helper
Unlike the openStream() function in other test classes, the one here takes a
performance mode to open the stream with, and tests exist which are identical
save for the mode they specify. However, it appears that the openStream()
implementation has never[1] actually paid attention to the performance mode,
meaning all tests run with the default one. Fix this by invoking
mBuilder.setPerformanceMode().

This fix does not appear to reveal any bugs. All tests in this file still pass
for me on Android 10 after the change.

[1] Since it was initially introduced in commit 95131751b47b ("oboe tests: test
new waitForStateChange behavior")