oboe: remove OpenSL ES direct links (#2014)
Use dlsym() to find the address of symbols like SL_IID_PLAY.
This allows Oboe to be run on a device that does not
have "libOpenSLES.so".

This fixes b/337360630

Fixed a deadlock that could occur if OpenSL ES was used when
the library was not present. Failing to link the OpenSL ES engine
would call close(), which tried to lock mLock again.

Use dlerror() to print info about link problems.

Add a state variable for better management of the dynamic linking.