OboeTester - prevent null crash with automation
I tried to prevent getResetCount() from being called when the stream is not ready.
But it was hard to track down the offending call.
So I just prevented the NULL crash.

I then added a dump of the frames read and written so we can see if recording is blocked
due to permissions.

Surprisingly the test semed to run OK after this patch.

Fixes #2201